+++ android/src/main/kotlin/com/adyen/checkout/flutter/utils/Constants.kt
@@ -6,8 +6,6 @@ class Constants {
             "FlutterFragmentActivity not used. Your activity needs to inherit from FlutterFragmentActivity."
         const val UNKNOWN_PAYMENT_METHOD_TYPE_ERROR_MESSAGE = "Unknown payment method type."
         const val ADYEN_LOG_TAG = "ADYEN_CHECKOUT"
-        const val GOOGLE_PAY_SESSION_COMPONENT_KEY = "GOOGLE_PAY_SESSION_COMPONENT"
-        const val GOOGLE_PAY_ADVANCED_COMPONENT_KEY = "GOOGLE_PAY_ADVANCED_COMPONENT"
         const val INSTANT_SESSION_COMPONENT_KEY = "INSTANT_SESSION_COMPONENT"
         const val INSTANT_ADVANCED_COMPONENT_KEY = "INSTANT_ADVANCED_COMPONENT"
         const val SDK_PAYMENT_CANCELED_IDENTIFIER = "Payment canceled"