+++ lib/adyen_checkout.dart
@@ -50,10 +50,6 @@ export 'src/components/apple_pay/adyen_apple_pay_component.dart';
 export 'src/components/apple_pay/model/apple_pay_button_style.dart';
 export 'src/components/apple_pay/model/apple_pay_button_theme.dart';
 export 'src/components/apple_pay/model/apple_pay_button_type.dart';
-export 'src/components/google_pay/adyen_google_pay_component.dart';
-export 'src/components/google_pay/model/google_pay_button_style.dart';
-export 'src/components/google_pay/model/google_pay_button_theme.dart';
-export 'src/components/google_pay/model/google_pay_button_type.dart';
 export 'src/components/instant/model/instant_component_configuration.dart';
 export 'src/drop_in/model/drop_in_configuration.dart';
 export 'src/generated/platform_api.g.dart'