+++ lib/src/components/google_pay/model/google_pay_button_theme.dart
@@ -1,4 +1,0 @@
-enum GooglePayButtonTheme {
-  dark,
-  light,
-}